// AngularApp/src/app/models/employee.ts
export class Employee {
    employee_id : number;
    department: string;
    salary: number;
    job_level: string;
    gender: string;
    new_hire: string;
    vacation_days_taken: number;
    engagement: number;
    rating: number;
    accident_type: string;
    location: string;
    overtime_hours: number;
    year: number;
}